+++ CHANGELOG.md
@@ -24,6 +24,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Fixed issue that the ORF start site position cannot be interpreted when checking whether it is at least one loop away, because it can be off by 1 when converting the location from the gene coordinate to amino acid. #630
 
+- Fixed issue that the 'CHROM' attribute of GVF metadata not read in correctly. #629
+
+- Fixed issue that when a frameshift insertion is on a alt splice frameshift substitution (or insertion), the node became disconnected after aligning the variant bubble. #635
+
+- Fixed issue that when getting the stop altering mutations, location comparison was done incorrectly by 1. #636
+
+- Node in circRNA missing downstream stop lost mutation called as variant peptide incorrectly. #637
+
 ## [0.10.1] - 2022-11-2
 
 ### Fixed